+++ test/spec/cinema6.db.ut.js
@@ -49,7 +49,7 @@
 
                         adapter = this;
                     });
-                AdapterInjectable = ['config', '$q', AdapterConstructor];
+                AdapterInjectable = ['config', '$q', 'cinema6', AdapterConstructor];
 
                 module('c6.ui', function($injector) {
                     cinema6Provider = $injector.get('cinema6Provider');
@@ -74,10 +74,13 @@
                         inject(function($injector) {
                             cinema6 = $injector.get('cinema6');
                         });
+
+                        // Make a request to instantiate the adapter
+                        cinema6.db.findAll('foo');
                     });
 
                     it('should provide that config object to the adapter', function() {
-                        expect(AdapterConstructor).toHaveBeenCalledWith(AdapterInjectable.config, jasmine.any(Object));
+                        expect(AdapterConstructor).toHaveBeenCalledWith(AdapterInjectable.config, jasmine.any(Object), cinema6);
                     });
                 });
 
@@ -86,10 +89,13 @@
                         inject(function($injector) {
                             cinema6 = $injector.get('cinema6');
                         });
+
+                        // Make a request to instantiate the adapter
+                        cinema6.db.findAll('foo');
                     });
 
                     it('should provide an empty object', function() {
-                        expect(AdapterConstructor).toHaveBeenCalledWith({}, jasmine.any(Object));
+                        expect(AdapterConstructor).toHaveBeenCalledWith({}, jasmine.any(Object), cinema6);
                     });
                 });
             });
@@ -375,6 +381,10 @@
                         expect(newResult).toEqual(jasmine.objectContaining(newModel));
                         expect(newResult).toBe(result);
                     });
+
+                    it('should be a shallow copy', function() {
+                        expect(result.items).toBe(model.items);
+                    });
                 });
 
                 describe('create(type, data)', function() {
@@ -411,6 +421,10 @@
                         expect(result.save).toEqual(jasmine.any(Function));
                         expect(result.erase).toEqual(jasmine.any(Function));
                     });
+
+                    it('should be a shallow copy', function() {
+                        expect(result.data).toBe(data.data);
+                    });
                 });
 
                 describe('find(type, id)', function() {
@@ -422,7 +436,8 @@
                             {
                                 id: 'e-2ff054584731c6',
                                 type: 'minireel',
-                                user: 'u-38b61e71b25d1e'
+                                user: 'u-38b61e71b25d1e',
+                                data: {}
                             }
                         ];
 
@@ -445,6 +460,14 @@
                         });
                         expect(findSpy).toHaveBeenCalledWith(jasmine.objectContaining(result[0]));
                         assertIsDBModel(findSpy.mostRecentCall.args[0]);
+                    });
+
+                    it('should be a shallow copy', function() {
+                        $rootScope.$apply(function() {
+                            adapter._deferreds.find.resolve(result);
+                        });
+
+                        expect(findSpy.mostRecentCall.args[0].data).toBe(result[0].data);
                     });
 
                     it('should cache the item by type and id', function() {